LayoutTests/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/get-persistendeviceid-from-pointer-event.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_equals: deviceId is 0 expected (number) 0 but got (undefined) undefined

Harness Error (FAIL), message = Error: assert_equals: deviceId is 0 expected (number) 0 but got (undefined) undefined

FAIL PointerEvent.persistentDeviceId assert_true: expected true got false
